есть архитектура, правка багов

+++ API/ImageService/Consumers/ImageConsumer.cs
@@ -5,6 +5,7 @@ using Common.Contracts.Processing;
 using ImageService.Data;
 using MassTransit;
 using Microsoft.EntityFrameworkCore;
+using ImageService.Entities;
 
 namespace ImageService.Consumers;
 
@@ -24,12 +25,26 @@ public class ImageConsumer : IConsumer<DataForProcessingServicesList<AuctionItem
     {
         var typedItem = JsonSerializer.Deserialize<AuctionItem>(context.Message.DataObjects[0].Data);
         var correlationId = context.Message.CorrelationId;
-        var item = await _context.Images.FirstOrDefaultAsync(p => p.AuctionId == typedItem.AuctionId);
-        if (item != null)
+        if (context.Message.DataObjects[0].CRUD == CRUD.Delete)
         {
-            _context.Images.Remove(item);
-            await _context.SaveChangesAsync();
+            var item = await _context.Images.FirstOrDefaultAsync(p => p.AuctionId == typedItem.AuctionId);
+            if (item != null)
+            {
+                _context.Images.Remove(item);
+            }
         }
+        if (context.Message.DataObjects[0].CRUD == CRUD.Create)
+        {
+            await _context.Images.AddAsync(new ImageItem
+            {
+                AuctionId = typedItem.AuctionId,
+                Image = Convert.FromBase64String(typedItem.Winner
+                         .Replace("data:image/png;base64,", "")
+                         .Replace("data:image/jpeg;base64,", "")
+                         .Replace("data:image/jpg;base64,", ""))
+            });
+        }
+        await _context.SaveChangesAsync();
         var sendObject = Assembly.LoadFrom(Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location) +
             _configuration["CommonAssembly"]).CreateInstance(context.Message.CallBackType);
         sendObject.GetType().GetProperty("CorrelationId").SetValue(sendObject, correlationId);
